+++ client/components/basic/Logo.js
@@ -3,6 +3,7 @@ import React, { useState } from 'react';
 
 function Logo({ src = 'images/logo/logo.svg', ...props }) {
 	const [isLoaded, setLoaded] = useState(false);
+	const isPlaceholderVisible = !src || !isLoaded;
 
 	const handleLoad = () => {
 		setLoaded(true);
@@ -13,7 +14,7 @@ function Logo({ src = 'images/logo/logo.svg', ...props }) {
 	};
 
 	return <>
-		{!isLoaded && <Box is='svg' viewBox='0 0 1000 182' xmlns='http://www.w3.org/2000/svg' {...props}>
+		{isPlaceholderVisible && <Box {...props} is='svg' xmlns='http://www.w3.org/2000/svg' viewBox='0 0 1000 182'>
 			<g fill='none' fillRule='evenodd'>
 				<g fill='#9ea2a8' fillRule='nonzero' transform='translate(245.154709 29.581775)'>
 					<path d='m62.5717573 58.8485839c0 10.1578288-3.791296 16.8288535-11.0711519 20.01071l10.4656798 39.7212001c.4541041 1.822594-.454104 2.727981-2.1238826 2.727981h-15.7753863c-1.5160454 0-2.2728855-.756459-2.5779867-2.122813l-10.1629438-38.5061381h-10.4656799v38.2059181c0 1.517646-.9082081 2.423033-2.4266186 2.423033h-15.77538636c-1.51604538 0-2.42661865-.910115-2.42661865-2.423033v-116.43561996c0-1.51528236.91057327-2.42539734 2.42661865-2.42539734h38.07142346c14.2569758 0 21.8419329 7.58113963 21.8419329 21.83094zm-26.6951702 1.817866c3.942664 0 6.0665466-2.1228136 6.0665466-6.0634933v-28.5019646c0-3.9406797-2.1238826-6.0611294-6.0665466-6.0611294h-15.0161812v40.6289513z'/>
@@ -33,7 +34,7 @@ function Logo({ src = 'images/logo/logo.svg', ...props }) {
 				<path d='m65.7703624 103.026182c-6.6534511 0-12.047139-5.4803639-12.047139-12.2407349 0-6.7603709 5.3936879-12.2407344 12.047139-12.2407344 6.6534512 0 12.047139 5.4803635 12.047139 12.2407344 0 6.760371-5.3936878 12.2407349-12.047139 12.2407349zm38.7859106 0c-6.6534508 0-12.0471386-5.4803639-12.0471386-12.2407349 0-6.7603709 5.3936878-12.2407344 12.0471386-12.2407344 6.653452 0 12.047139 5.4803635 12.047139 12.2407344 0 6.760371-5.393687 12.2407349-12.047139 12.2407349zm38.785911 0c-6.653451 0-12.047139-5.4803639-12.047139-12.2407349 0-6.7603709 5.393688-12.2407344 12.047139-12.2407344 6.653452 0 12.047139 5.4803635 12.047139 12.2407344 0 6.760371-5.393687 12.2407349-12.047139 12.2407349z' fill='#db2323' fillRule='nonzero'/>
 			</g>
 		</Box>}
-		<Box is='img' src={src} hidden={!isLoaded} onLoad={handleLoad} onError={handleError} {...props} />
+		{src && <Box {...props} is='img' src={src} hidden={!isLoaded} onLoad={handleLoad} onError={handleError} />}
 	</>;
 }
